added FileManager modularity, NavBar, Grid and Listview done

+++ C78Engine/src/C78E/ImGui/TextureHandle.h
@@ -7,15 +7,15 @@ namespace C78E {
 
 	class TextureHandle {
 	public:
-		TextureHandle(uint32_t handle) : m_RendererID(handle) {}
-		TextureHandle(ImTextureID handle) : m_RendererID((uint32_t)handle) {}
+		TextureHandle(uint32_t handle) : m_RendererID(static_cast<size_t>(handle)) {}
+		TextureHandle(ImTextureID handle) : m_RendererID((size_t)handle) {}
 		TextureHandle(const TextureHandle& other) : m_RendererID(other.m_RendererID) {}
 		~TextureHandle() = default;
-
-		operator uint32_t() const { return m_RendererID; }
+		
+		operator uint32_t() const { return static_cast<uint32_t>(m_RendererID); }
 		operator ImTextureID() const { return (ImTextureID)m_RendererID; }
 	private:
-		uint32_t m_RendererID = 0;
+		size_t m_RendererID = 0;
 	};
 
 }
